# openapi/v1.yaml
openapi: "3.0.3"
info:
  title: EdgeFabric Controller API
  description: |
    REST API for the EdgeFabric distributed edge networking platform.
    Provides multi-tenant management, user authentication (password + TOTP 2FA),
    API key management, fleet management (nodes, node groups, SSH keys),
    audit logging, and status dashboard.
  version: "1.1.0"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8443
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/nodes:
    get:
      tags: [Nodes]
      summary: List nodes
      description: Non-superusers see only their tenant's nodes. SuperUsers see all nodes.
      parameters:
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated list of nodes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Nodes]
      summary: Create a node
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, hostname, public_ip]
              properties:
                name:
                  type: string
                hostname:
                  type: string
                public_ip:
                  type: string
                  format: ipv4
                region:
                  type: string
                provider:
                  type: string
                ssh_port:
                  type: integer
                  default: 22
                ssh_user:
                  type: string
                  default: root
      responses:
        "201":
          description: Node created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse_Node"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Node already exists
  /api/v1/nodes/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Nodes]
      summary: Get a node by ID
      description: Enforces tenant isolation for non-superuser callers.
      responses:
        "200":
          description: Node details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse_Node"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Nodes]
      summary: Update a node
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                tenant_id:
                  type: string
                  format: uuid
                status:
                  $ref: "#/components/schemas/NodeStatus"
                region:
                  type: string
      responses:
        "200":
          description: Node updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse_Node"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Nodes]
      summary: Delete a node
      responses:
        "204":
          description: Node deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/nodes/{id}/heartbeat:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      tags: [Nodes]
      summary: Record a node heartbeat
      description: Updates the node's last heartbeat timestamp and sets status to online.
      responses:
        "200":
          description: Heartbeat recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                        example: ok
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Session token (from /auth/login) or API key (prefixed with ef_).
        Example: `Authorization: Bearer <token>` or `Authorization: Bearer ef_<key>`
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  schemas:
...
    NodeStatus:
      type: string
      enum: [pending, enrolling, online, offline, error, decommissioned]
    NodeCapability:
      type: string
      enum: [bgp, dns, cdn, route]
    Node:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        hostname:
          type: string
        public_ip:
          type: string
        wireguard_ip:
          type: string
        status:
          $ref: "#/components/schemas/NodeStatus"
        region:
          type: string
        provider:
          type: string
        ssh_port:
          type: integer
        ssh_user:
          type: string
        ssh_key_id:
          type: string
          format: uuid
          nullable: true
        binary_version:
          type: string
        last_heartbeat:
          type: string
          format: date-time
          nullable: true
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/NodeCapability"
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    DataResponse_Node:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Node"
...
    NodeListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Node"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"